feat: shared environment support added
