+++ packages/aws-cdk-lib/aws-dynamodb/lib/global-table.ts
@@ -1,139 +1,15 @@
-import * as appscaling from '../../aws-applicationautoscaling';
-import * as cloudwatch from '../../aws-cloudwatch';
 import * as iam from '../../aws-iam';
 import * as kinesis from '../../aws-kinesis';
 import * as kms from '../../aws-kms';
-import {
-  ArnFormat,
-  Aws, CfnCondition, CfnCustomResource, CfnResource, CustomResource, Duration,
-  Fn, IResource, Lazy, Names, RemovalPolicy, Resource, Stack, Token,
-} from '../../core';
+import { ArnFormat, Lazy, RemovalPolicy, Stack, Token } from "../../core";
 import { Construct } from 'constructs';
-import { DynamoDBMetrics } from './dynamodb-canned-metrics.generated';
 import { CfnGlobalTable, CfnTable, CfnTableProps } from './dynamodb.generated';
-import * as perms from './perms';
-import { ReplicaProvider } from './replica-provider';
-import { EnableScalingProps, IScalableTableAttribute } from './scalable-attribute-api';
-import { ScalableTableAttribute } from './scalable-table-attribute';
+import {
+    Attribute, BillingMode, HASH_KEY_TYPE, ITable, LocalSecondaryIndexProps, MAX_LOCAL_SECONDARY_INDEX_COUNT,
+    ProjectionType, RANGE_KEY_TYPE, SecondaryIndexProps, StreamViewType, TableAttributes, TableBase, TableClass, TableEncryption
+} from "./share";
 
-const HASH_KEY_TYPE = 'HASH';
-const RANGE_KEY_TYPE = 'RANGE';
-
-
-// https://docs.aws.amazon.com/amazondynamodb/latest/developerguide/Limits.html#limits-secondary-indexes
-const MAX_LOCAL_SECONDARY_INDEX_COUNT = 5;
-
-/**
- * Options for configuring a system errors metric that considers multiple operations.
- */
-export interface SystemErrorsForOperationsMetricOptions extends cloudwatch.MetricOptions {
-
-    /**
-     * The operations to apply the metric to.
-     *
-     * @default - All operations available by DynamoDB tables will be considered.
-     */
-    readonly operations?: Operation[];
-
-}
-
-/**
- * Options for configuring metrics that considers multiple operations.
- */
-export interface OperationsMetricOptions extends SystemErrorsForOperationsMetricOptions { }
-
-/**
- * Supported DynamoDB table operations.
- */
-export enum Operation {
-
-    /** GetItem */
-    GET_ITEM = 'GetItem',
-
-    /** BatchGetItem */
-    BATCH_GET_ITEM = 'BatchGetItem',
-
-    /** Scan */
-    SCAN = 'Scan',
-
-    /** Query */
-    QUERY = 'Query',
-
-    /** GetRecords */
-    GET_RECORDS = 'GetRecords',
-
-    /** PutItem */
-    PUT_ITEM = 'PutItem',
-
-    /** DeleteItem */
-    DELETE_ITEM = 'DeleteItem',
-
-    /** UpdateItem */
-    UPDATE_ITEM = 'UpdateItem',
-
-    /** BatchWriteItem */
-    BATCH_WRITE_ITEM = 'BatchWriteItem',
-
-    /** TransactWriteItems */
-    TRANSACT_WRITE_ITEMS = 'TransactWriteItems',
-
-    /** TransactGetItems */
-    TRANSACT_GET_ITEMS = 'TransactGetItems',
-
-    /** ExecuteTransaction */
-    EXECUTE_TRANSACTION = 'ExecuteTransaction',
-
-    /** BatchExecuteStatement */
-    BATCH_EXECUTE_STATEMENT = 'BatchExecuteStatement',
-
-    /** ExecuteStatement */
-    EXECUTE_STATEMENT = 'ExecuteStatement',
-
-}
-
-/**
- * Represents an attribute for describing the key schema for the table
- * and indexes.
- */
-export interface Attribute {
-    /**
-     * The name of an attribute.
-     */
-    readonly name: string;
-
-    /**
-     * The data type of an attribute.
-     */
-    readonly type: AttributeType;
-}
-
-/**
- * What kind of server-side encryption to apply to this table.
- */
-export enum TableEncryption {
-    /**
-     * Server-side KMS encryption with a master key owned by AWS.
-     */
-    DEFAULT = 'AWS_OWNED',
-
-    /**
-     * Server-side KMS encryption with a customer master key managed by customer.
-     * If `encryptionKey` is specified, this key will be used, otherwise, one will be defined.
-     *
-     * > **NOTE**: if `encryptionKey` is not specified and the `Table` construct creates
-     * > a KMS key for you, the key will be created with default permissions. If you are using
-     * > CDKv2, these permissions will be sufficient to enable the key for use with DynamoDB tables.
-     * > If you are using CDKv1, make sure the feature flag `@aws-cdk/aws-kms:defaultKeyPolicies`
-     * > is set to `true` in your `cdk.json`.
-     */
-    CUSTOMER_MANAGED = 'CUSTOMER_MANAGED',
-
-    /**
-     * Server-side KMS encryption with a master key managed by AWS.
-     */
-    AWS_MANAGED = 'AWS_MANAGED',
-}
-
+// remove
 /**
  * Represents the table schema attributes.
  */
@@ -156,7 +32,7 @@ export interface SchemaOptions {
  *
  * Use `TableProps` for all table properties
  */
-export interface TableOptions extends SchemaOptions {
+export interface TableOptions {
     /**
      * The read capacity for the table. Careful if you add Global Secondary Indexes, as
      * those will share the table's provisioned throughput.
@@ -257,36 +133,6 @@ export interface TableOptions extends SchemaOptions {
     readonly removalPolicy?: RemovalPolicy;
 
     /**
-     * Regions where replica tables will be created
-     *
-     * @default - no replica tables are created
-     */
-    readonly replicationRegions?: string[];
-
-    /**
-     * The timeout for a table replication operation in a single region.
-     *
-     * @default Duration.minutes(30)
-     */
-    readonly replicationTimeout?: Duration;
-
-    /**
-     * Indicates whether CloudFormation stack waits for replication to finish.
-     * If set to false, the CloudFormation resource will mark the resource as
-     * created and replication will be completed asynchronously. This property is
-     * ignored if replicationRegions property is not set.
-     *
-     * DO NOT UNSET this property if adding/removing multiple replicationRegions
-     * in one deployment, as CloudFormation only supports one region replication
-     * at a time. CDK overcomes this limitation by waiting for replication to
-     * finish before starting new replicationRegion.
-     *
-     * @see https://docs.aws.amazon.com/AWSCloudFormation/latest/UserGuide/aws-resource-dynamodb-globaltable.html#cfn-dynamodb-globaltable-replicas
-     * @default true
-     */
-    readonly waitForReplicationToFinish?: boolean;
-
-    /**
      * Whether CloudWatch contributor insights is enabled.
      *
      * @default false
@@ -298,6 +144,18 @@ export interface TableOptions extends SchemaOptions {
  * Properties for a DynamoDB Table
  */
 export interface TableProps extends TableOptions {
+    /**
+     * Partition key attribute definition.
+     */
+    readonly partitionKey: Attribute;
+
+    /**
+     * Sort key attribute definition.
+     *
+     * @default no sort key
+     */
+    readonly sortKey?: Attribute;
+
     /**
      * Enforces a particular physical table name.
      * @default <generated>
@@ -318,10 +176,10 @@ export interface TableProps extends TableOptions {
      *
      * @default - The list contains one element, the region where the stack defining the global table is deployed.
      */
-    readonly replicas?: ReplicaProps
+    readonly replicas?: ReplicaProps[]
 
     /**
-     * capacity
+     * Specifies the capacities of
      */
     readonly capacity?: Capacity
 
@@ -338,24 +196,6 @@ export interface TableProps extends TableOptions {
      */
     readonly readProvisionedThroughputSettings?: CfnGlobalTable.ReadProvisionedThroughputSettingsProperty;
 }
-
-// interface writeProvisionedThroughputSettingsProps {
-//       /**
-//    * Minimum capacity to scale to
-//    */
-//   readonly minCapacity: number;
-
-//   /**
-//    * Maximum capacity to scale to
-//    */
-//   readonly maxCapacity: number;
-
-//   readonly targetTrackingScalingPolicyConfiguration: targetTrackingScalingPolicyConfigurationProps
-// }
-
-// interface targetTrackingScalingPolicyConfigurationProps {
-//     targetValue: number
-// }
 
 /**
  * Properties for Replication
@@ -414,780 +254,26 @@ export interface ReplicaProps {
 }
 
 /**
- * Properties for a secondary index
- */
-export interface SecondaryIndexProps {
-    /**
-     * The name of the secondary index.
-     */
-    readonly indexName: string;
-
-    /**
-     * The set of attributes that are projected into the secondary index.
-     * @default ALL
-     */
-    readonly projectionType?: ProjectionType;
-
-    /**
-     * The non-key attributes that are projected into the secondary index.
-     * @default - No additional attributes
-     */
-    readonly nonKeyAttributes?: string[];
-}
-
-/**
  * Properties for a global secondary index
  */
-export interface GlobalSecondaryIndexProps extends SecondaryIndexProps, SchemaOptions {
+export interface GlobalSecondaryIndexProps extends SecondaryIndexProps {
     /**
-     * The read capacity for the global secondary index.
-     *
-     * Can only be provided if table billingMode is Provisioned or undefined.
-     *
-     * @default 5
+     * Partition key attribute definition.
      */
-    readonly readCapacity?: number;
+    readonly partitionKey: Attribute;
 
     /**
-     * The write capacity for the global secondary index.
+     * Sort key attribute definition.
      *
-     * Can only be provided if table billingMode is Provisioned or undefined.
-     *
-     * @default 5
+     * @default no sort key
      */
-    readonly writeCapacity?: number;
+    readonly sortKey?: Attribute;
 
-    /**
-     * Specifies an auto scaling policy for write capacity.
-     * This policy will be applied to all replicas. This setting must be specified
-     * if `BillingMode` is set to `PROVISIONED`.
-     */
-    writeProvisionedThroughputSettings?: CfnGlobalTable.WriteProvisionedThroughputSettingsProperty
-
-    /**
-     * Allows you to specify the read capacity settings for a replica global secondary index
-     * when the BillingMode is set to PROVISIONED. You must specify a value for either `ReadCapacityUnits`
-     * or `ReadCapacityAutoScalingSettings`, but not both. You can switch between fixed capacity and auto scaling.
-     */
-    readProvisionedThroughputSettings?: CfnGlobalTable.ReadProvisionedThroughputSettingsProperty
+    readonly capacity?: Capacity
 }
 
 /**
- * Properties for a local secondary index
- */
-export interface LocalSecondaryIndexProps extends SecondaryIndexProps {
-    /**
-     * The attribute of a sort key for the local secondary index.
-     */
-    readonly sortKey: Attribute;
-}
-
-/**
- * An interface that represents a DynamoDB Table - either created with the CDK, or an existing one.
- */
-export interface ITable extends IResource {
-    /**
-     * Arn of the dynamodb table.
-     *
-     * @attribute
-     */
-    readonly tableArn: string;
-
-    /**
-     * Table name of the dynamodb table.
-     *
-     * @attribute
-     */
-    readonly tableName: string;
-
-    /**
-     * ARN of the table's stream, if there is one.
-     *
-     * @attribute
-     */
-    readonly tableStreamArn?: string;
-
-    /**
-     *
-     * Optional KMS encryption key associated with this table.
-     */
-    readonly encryptionKey?: kms.IKey;
-
-    /**
-     * Adds an IAM policy statement associated with this table to an IAM
-     * principal's policy.
-     *
-     * If `encryptionKey` is present, appropriate grants to the key needs to be added
-     * separately using the `table.encryptionKey.grant*` methods.
-     *
-     * @param grantee The principal (no-op if undefined)
-     * @param actions The set of actions to allow (i.e. "dynamodb:PutItem", "dynamodb:GetItem", ...)
-     */
-    grant(grantee: iam.IGrantable, ...actions: string[]): iam.Grant;
-
-    /**
-     * Adds an IAM policy statement associated with this table's stream to an
-     * IAM principal's policy.
-     *
-     * If `encryptionKey` is present, appropriate grants to the key needs to be added
-     * separately using the `table.encryptionKey.grant*` methods.
-     *
-     * @param grantee The principal (no-op if undefined)
-     * @param actions The set of actions to allow (i.e. "dynamodb:DescribeStream", "dynamodb:GetRecords", ...)
-     */
-    grantStream(grantee: iam.IGrantable, ...actions: string[]): iam.Grant;
-
-    /**
-     * Permits an IAM principal all data read operations from this table:
-     * BatchGetItem, GetRecords, GetShardIterator, Query, GetItem, Scan.
-     *
-     * Appropriate grants will also be added to the customer-managed KMS key
-     * if one was configured.
-     *
-     * @param grantee The principal to grant access to
-     */
-    grantReadData(grantee: iam.IGrantable): iam.Grant;
-
-    /**
-     * Permits an IAM Principal to list streams attached to current dynamodb table.
-     *
-     * @param grantee The principal (no-op if undefined)
-     */
-    grantTableListStreams(grantee: iam.IGrantable): iam.Grant;
-
-    /**
-     * Permits an IAM principal all stream data read operations for this
-     * table's stream:
-     * DescribeStream, GetRecords, GetShardIterator, ListStreams.
-     *
-     * Appropriate grants will also be added to the customer-managed KMS key
-     * if one was configured.
-     *
-     * @param grantee The principal to grant access to
-     */
-    grantStreamRead(grantee: iam.IGrantable): iam.Grant;
-
-    /**
-     * Permits an IAM principal all data write operations to this table:
-     * BatchWriteItem, PutItem, UpdateItem, DeleteItem.
-     *
-     * Appropriate grants will also be added to the customer-managed KMS key
-     * if one was configured.
-     *
-     * @param grantee The principal to grant access to
-     */
-    grantWriteData(grantee: iam.IGrantable): iam.Grant;
-
-    /**
-     * Permits an IAM principal to all data read/write operations to this table.
-     * BatchGetItem, GetRecords, GetShardIterator, Query, GetItem, Scan,
-     * BatchWriteItem, PutItem, UpdateItem, DeleteItem
-     *
-     * Appropriate grants will also be added to the customer-managed KMS key
-     * if one was configured.
-     *
-     * @param grantee The principal to grant access to
-     */
-    grantReadWriteData(grantee: iam.IGrantable): iam.Grant;
-
-    /**
-     * Permits all DynamoDB operations ("dynamodb:*") to an IAM principal.
-     *
-     * Appropriate grants will also be added to the customer-managed KMS key
-     * if one was configured.
-     *
-     * @param grantee The principal to grant access to
-     */
-    grantFullAccess(grantee: iam.IGrantable): iam.Grant;
-
-    /**
-     * Metric for the number of Errors executing all Lambdas
-     */
-    metric(metricName: string, props?: cloudwatch.MetricOptions): cloudwatch.Metric;
-
-    /**
-     * Metric for the consumed read capacity units
-     *
-     * @param props properties of a metric
-     */
-    metricConsumedReadCapacityUnits(props?: cloudwatch.MetricOptions): cloudwatch.Metric;
-
-    /**
-     * Metric for the consumed write capacity units
-     *
-     * @param props properties of a metric
-     */
-    metricConsumedWriteCapacityUnits(props?: cloudwatch.MetricOptions): cloudwatch.Metric;
-
-    /**
-     * Metric for the system errors
-     *
-     * @param props properties of a metric
-     *
-     * @deprecated use `metricSystemErrorsForOperations`
-     */
-    metricSystemErrors(props?: cloudwatch.MetricOptions): cloudwatch.Metric;
-
-    /**
-     * Metric for the system errors this table
-     *
-     * @param props properties of a metric
-     *
-     */
-    metricSystemErrorsForOperations(props?: SystemErrorsForOperationsMetricOptions): cloudwatch.IMetric;
-
-    /**
-     * Metric for the user errors
-     *
-     * @param props properties of a metric
-     */
-    metricUserErrors(props?: cloudwatch.MetricOptions): cloudwatch.Metric;
-
-    /**
-     * Metric for the conditional check failed requests
-     *
-     * @param props properties of a metric
-     */
-    metricConditionalCheckFailedRequests(props?: cloudwatch.MetricOptions): cloudwatch.Metric;
-
-    /**
-     * Metric for throttled requests
-     *
-     * @param props properties of a metric
-     *
-     * @deprecated use `metricThrottledRequestsForOperations`
-     */
-    metricThrottledRequests(props?: cloudwatch.MetricOptions): cloudwatch.Metric;
-
-    /**
-     * Metric for throttled requests
-     *
-     * @param props properties of a metric
-     *
-     */
-    metricThrottledRequestsForOperations(props?: OperationsMetricOptions): cloudwatch.IMetric;
-
-    /**
-     * Metric for the successful request latency
-     *
-     * @param props properties of a metric
-     *
-     */
-    metricSuccessfulRequestLatency(props?: cloudwatch.MetricOptions): cloudwatch.Metric;
-}
-
-/**
- * Reference to a dynamodb table.
- */
-export interface TableAttributes {
-    /**
-     * The ARN of the dynamodb table.
-     * One of this, or `tableName`, is required.
-     *
-     * @default - no table arn
-     */
-    readonly tableArn?: string;
-
-    /**
-     * The table name of the dynamodb table.
-     * One of this, or `tableArn`, is required.
-     *
-     * @default - no table name
-     */
-    readonly tableName?: string;
-
-    /**
-     * The ARN of the table's stream.
-     *
-     * @default - no table stream
-     */
-    readonly tableStreamArn?: string;
-
-    /**
-     * KMS encryption key, if this table uses a customer-managed encryption key.
-     *
-     * @default - no key
-     */
-    readonly encryptionKey?: kms.IKey;
-
-    /**
-     * The name of the global indexes set for this Table.
-     * Note that you need to set either this property,
-     * or `localIndexes`,
-     * if you want methods like grantReadData()
-     * to grant permissions for indexes as well as the table itself.
-     *
-     * @default - no global indexes
-     */
-    readonly globalIndexes?: string[];
-
-    /**
-     * The name of the local indexes set for this Table.
-     * Note that you need to set either this property,
-     * or `globalIndexes`,
-     * if you want methods like grantReadData()
-     * to grant permissions for indexes as well as the table itself.
-     *
-     * @default - no local indexes
-     */
-    readonly localIndexes?: string[];
-
-    /**
-     * If set to true, grant methods always grant permissions for all indexes.
-     * If false is provided, grant methods grant the permissions
-     * only when `globalIndexes` or `localIndexes` is specified.
-     *
-     * @default - false
-     */
-    readonly grantIndexPermissions?: boolean;
-}
-
-abstract class TableBase extends Resource implements ITable {
-    /**
-     * @attribute
-     */
-    public abstract readonly tableArn: string;
-
-    /**
-     * @attribute
-     */
-    public abstract readonly tableName: string;
-
-    /**
-     * @attribute
-     */
-    public abstract readonly tableStreamArn?: string;
-
-    /**
-     * KMS encryption key, if this table uses a customer-managed encryption key.
-     */
-    public abstract readonly encryptionKey?: kms.IKey;
-
-    protected readonly regionalArns = new Array<string>();
-
-    /**
-     * Adds an IAM policy statement associated with this table to an IAM
-     * principal's policy.
-     *
-     * If `encryptionKey` is present, appropriate grants to the key needs to be added
-     * separately using the `table.encryptionKey.grant*` methods.
-     *
-     * @param grantee The principal (no-op if undefined)
-     * @param actions The set of actions to allow (i.e. "dynamodb:PutItem", "dynamodb:GetItem", ...)
-     */
-    public grant(grantee: iam.IGrantable, ...actions: string[]): iam.Grant {
-        return iam.Grant.addToPrincipal({
-            grantee,
-            actions,
-            resourceArns: [
-                this.tableArn,
-                Lazy.string({ produce: () => this.hasIndex ? `${this.tableArn}/index/*` : Aws.NO_VALUE }),
-                ...this.regionalArns,
-                ...this.regionalArns.map(arn => Lazy.string({
-                    produce: () => this.hasIndex ? `${arn}/index/*` : Aws.NO_VALUE,
-                })),
-            ],
-            scope: this,
-        });
-    }
-
-    /**
-     * Adds an IAM policy statement associated with this table's stream to an
-     * IAM principal's policy.
-     *
-     * If `encryptionKey` is present, appropriate grants to the key needs to be added
-     * separately using the `table.encryptionKey.grant*` methods.
-     *
-     * @param grantee The principal (no-op if undefined)
-     * @param actions The set of actions to allow (i.e. "dynamodb:DescribeStream", "dynamodb:GetRecords", ...)
-     */
-    public grantStream(grantee: iam.IGrantable, ...actions: string[]): iam.Grant {
-        if (!this.tableStreamArn) {
-            throw new Error(`DynamoDB Streams must be enabled on the table ${this.node.path}`);
-        }
-
-        return iam.Grant.addToPrincipal({
-            grantee,
-            actions,
-            resourceArns: [this.tableStreamArn],
-            scope: this,
-        });
-    }
-
-    /**
-     * Permits an IAM principal all data read operations from this table:
-     * BatchGetItem, GetRecords, GetShardIterator, Query, GetItem, Scan, DescribeTable.
-     *
-     * Appropriate grants will also be added to the customer-managed KMS key
-     * if one was configured.
-     *
-     * @param grantee The principal to grant access to
-     */
-    public grantReadData(grantee: iam.IGrantable): iam.Grant {
-        const tableActions = perms.READ_DATA_ACTIONS.concat(perms.DESCRIBE_TABLE);
-        return this.combinedGrant(grantee, { keyActions: perms.KEY_READ_ACTIONS, tableActions });
-    }
-
-    /**
-     * Permits an IAM Principal to list streams attached to current dynamodb table.
-     *
-     * @param grantee The principal (no-op if undefined)
-     */
-    public grantTableListStreams(grantee: iam.IGrantable): iam.Grant {
-        if (!this.tableStreamArn) {
-            throw new Error(`DynamoDB Streams must be enabled on the table ${this.node.path}`);
-        }
-
-        return iam.Grant.addToPrincipal({
-            grantee,
-            actions: ['dynamodb:ListStreams'],
-            resourceArns: ['*'],
-        });
-    }
-
-    /**
-     * Permits an IAM principal all stream data read operations for this
-     * table's stream:
-     * DescribeStream, GetRecords, GetShardIterator, ListStreams.
-     *
-     * Appropriate grants will also be added to the customer-managed KMS key
-     * if one was configured.
-     *
-     * @param grantee The principal to grant access to
-     */
-    public grantStreamRead(grantee: iam.IGrantable): iam.Grant {
-        this.grantTableListStreams(grantee);
-        return this.combinedGrant(grantee, { keyActions: perms.KEY_READ_ACTIONS, streamActions: perms.READ_STREAM_DATA_ACTIONS });
-    }
-
-    /**
-     * Permits an IAM principal all data write operations to this table:
-     * BatchWriteItem, PutItem, UpdateItem, DeleteItem, DescribeTable.
-     *
-     * Appropriate grants will also be added to the customer-managed KMS key
-     * if one was configured.
-     *
-     * @param grantee The principal to grant access to
-     */
-    public grantWriteData(grantee: iam.IGrantable): iam.Grant {
-        const tableActions = perms.WRITE_DATA_ACTIONS.concat(perms.DESCRIBE_TABLE);
-        const keyActions = perms.KEY_READ_ACTIONS.concat(perms.KEY_WRITE_ACTIONS);
-        return this.combinedGrant(grantee, { keyActions, tableActions });
-    }
-
-    /**
-     * Permits an IAM principal to all data read/write operations to this table.
-     * BatchGetItem, GetRecords, GetShardIterator, Query, GetItem, Scan,
-     * BatchWriteItem, PutItem, UpdateItem, DeleteItem, DescribeTable
-     *
-     * Appropriate grants will also be added to the customer-managed KMS key
-     * if one was configured.
-     *
-     * @param grantee The principal to grant access to
-     */
-    public grantReadWriteData(grantee: iam.IGrantable): iam.Grant {
-        const tableActions = perms.READ_DATA_ACTIONS.concat(perms.WRITE_DATA_ACTIONS).concat(perms.DESCRIBE_TABLE);
-        const keyActions = perms.KEY_READ_ACTIONS.concat(perms.KEY_WRITE_ACTIONS);
-        return this.combinedGrant(grantee, { keyActions, tableActions });
-    }
-
-    /**
-     * Permits all DynamoDB operations ("dynamodb:*") to an IAM principal.
-     *
-     * Appropriate grants will also be added to the customer-managed KMS key
-     * if one was configured.
-     *
-     * @param grantee The principal to grant access to
-     */
-    public grantFullAccess(grantee: iam.IGrantable) {
-        const keyActions = perms.KEY_READ_ACTIONS.concat(perms.KEY_WRITE_ACTIONS);
-        return this.combinedGrant(grantee, { keyActions, tableActions: ['dynamodb:*'] });
-    }
-
-    /**
-     * Return the given named metric for this Table
-     *
-     * By default, the metric will be calculated as a sum over a period of 5 minutes.
-     * You can customize this by using the `statistic` and `period` properties.
-     */
-    public metric(metricName: string, props?: cloudwatch.MetricOptions): cloudwatch.Metric {
-        return new cloudwatch.Metric({
-            namespace: 'AWS/DynamoDB',
-            metricName,
-            dimensionsMap: {
-                TableName: this.tableName,
-            },
-            ...props,
-        }).attachTo(this);
-    }
-
-    /**
-     * Metric for the consumed read capacity units this table
-     *
-     * By default, the metric will be calculated as a sum over a period of 5 minutes.
-     * You can customize this by using the `statistic` and `period` properties.
-     */
-    public metricConsumedReadCapacityUnits(props?: cloudwatch.MetricOptions): cloudwatch.Metric {
-        return this.cannedMetric(DynamoDBMetrics.consumedReadCapacityUnitsSum, props);
-    }
-
-    /**
-     * Metric for the consumed write capacity units this table
-     *
-     * By default, the metric will be calculated as a sum over a period of 5 minutes.
-     * You can customize this by using the `statistic` and `period` properties.
-     */
-    public metricConsumedWriteCapacityUnits(props?: cloudwatch.MetricOptions): cloudwatch.Metric {
-        return this.cannedMetric(DynamoDBMetrics.consumedWriteCapacityUnitsSum, props);
-    }
-
-    /**
-     * Metric for the system errors this table
-     *
-     * @deprecated use `metricSystemErrorsForOperations`.
-     */
-    public metricSystemErrors(props?: cloudwatch.MetricOptions): cloudwatch.Metric {
-        if (!props?.dimensionsMap?.Operation && !props?.dimensionsMap?.Operation) {
-            // 'Operation' must be passed because its an operational metric.
-            throw new Error("'Operation' dimension must be passed for the 'SystemErrors' metric.");
-        }
-
-        const dimensionsMap = {
-            TableName: this.tableName,
-            ...props?.dimensionsMap ?? {},
-            ...props?.dimensionsMap ?? {},
-        };
-
-        return this.metric('SystemErrors', { statistic: 'sum', ...props, dimensionsMap });
-    }
-
-    /**
-     * Metric for the user errors. Note that this metric reports user errors across all
-     * the tables in the account and region the table resides in.
-     *
-     * By default, the metric will be calculated as a sum over a period of 5 minutes.
-     * You can customize this by using the `statistic` and `period` properties.
-     */
-    public metricUserErrors(props?: cloudwatch.MetricOptions): cloudwatch.Metric {
-        if (props?.dimensionsMap) {
-            throw new Error("'dimensions' is not supported for the 'UserErrors' metric");
-        }
-
-        // overriding 'dimensions' here because this metric is an account metric.
-        // see 'UserErrors' in https://docs.aws.amazon.com/amazondynamodb/latest/developerguide/metrics-dimensions.html
-        return this.metric('UserErrors', { statistic: 'sum', ...props, dimensionsMap: {} });
-    }
-
-    /**
-     * Metric for the conditional check failed requests this table
-     *
-     * By default, the metric will be calculated as a sum over a period of 5 minutes.
-     * You can customize this by using the `statistic` and `period` properties.
-     */
-    public metricConditionalCheckFailedRequests(props?: cloudwatch.MetricOptions): cloudwatch.Metric {
-        return this.metric('ConditionalCheckFailedRequests', { statistic: 'sum', ...props });
-    }
-
-    /**
-     * How many requests are throttled on this table
-     *
-     * Default: sum over 5 minutes
-     *
-     * @deprecated Do not use this function. It returns an invalid metric. Use `metricThrottledRequestsForOperation` instead.
-     */
-    public metricThrottledRequests(props?: cloudwatch.MetricOptions): cloudwatch.Metric {
-        return this.metric('ThrottledRequests', { statistic: 'sum', ...props });
-    }
-
-    /**
-     * Metric for the successful request latency this table.
-     *
-     * By default, the metric will be calculated as an average over a period of 5 minutes.
-     * You can customize this by using the `statistic` and `period` properties.
-     */
-    public metricSuccessfulRequestLatency(props?: cloudwatch.MetricOptions): cloudwatch.Metric {
-        if (!props?.dimensionsMap?.Operation && !props?.dimensionsMap?.Operation) {
-            throw new Error("'Operation' dimension must be passed for the 'SuccessfulRequestLatency' metric.");
-        }
-
-        const dimensionsMap = {
-            TableName: this.tableName,
-            Operation: props.dimensionsMap?.Operation ?? props.dimensionsMap?.Operation,
-        };
-
-        return new cloudwatch.Metric({
-            ...DynamoDBMetrics.successfulRequestLatencyAverage(dimensionsMap),
-            ...props,
-            dimensionsMap,
-        }).attachTo(this);
-    }
-
-    /**
-     * How many requests are throttled on this table, for the given operation
-     *
-     * Default: sum over 5 minutes
-     */
-    public metricThrottledRequestsForOperation(operation: string, props?: cloudwatch.MetricOptions): cloudwatch.Metric {
-        return new cloudwatch.Metric({
-            ...DynamoDBMetrics.throttledRequestsSum({ Operation: operation, TableName: this.tableName }),
-            ...props,
-        }).attachTo(this);
-    }
-
-    /**
-     * How many requests are throttled on this table.
-     *
-     * This will sum errors across all possible operations.
-     * Note that by default, each individual metric will be calculated as a sum over a period of 5 minutes.
-     * You can customize this by using the `statistic` and `period` properties.
-     */
-    public metricThrottledRequestsForOperations(props?: OperationsMetricOptions): cloudwatch.IMetric {
-        return this.sumMetricsForOperations('ThrottledRequests', 'Sum of throttled requests across all operations', props);
-    }
-
-    /**
-     * Metric for the system errors this table.
-     *
-     * This will sum errors across all possible operations.
-     * Note that by default, each individual metric will be calculated as a sum over a period of 5 minutes.
-     * You can customize this by using the `statistic` and `period` properties.
-     */
-    public metricSystemErrorsForOperations(props?: SystemErrorsForOperationsMetricOptions): cloudwatch.IMetric {
-        return this.sumMetricsForOperations('SystemErrors', 'Sum of errors across all operations', props);
-    }
-
-    /**
-     * Create a math expression for operations.
-     *
-     * @param metricName The metric name.
-     * @param expressionLabel Label for expression
-     * @param props operation list
-     */
-    private sumMetricsForOperations(metricName: string, expressionLabel: string, props?: OperationsMetricOptions): cloudwatch.IMetric {
-        if (props?.dimensionsMap?.Operation) {
-            throw new Error("The Operation dimension is not supported. Use the 'operations' property.");
-        }
-
-        const operations = props?.operations ?? Object.values(Operation);
-
-        const values = this.createMetricsForOperations(metricName, operations, { statistic: 'sum', ...props });
-
-        const sum = new cloudwatch.MathExpression({
-            expression: `${Object.keys(values).join(' + ')}`,
-            usingMetrics: { ...values },
-            color: props?.color,
-            label: expressionLabel,
-            period: props?.period,
-        });
-
-        return sum;
-    }
-
-    /**
-     * Create a map of metrics that can be used in a math expression.
-     *
-     * Using the return value of this function as the `usingMetrics` property in `cloudwatch.MathExpression` allows you to
-     * use the keys of this map as metric names inside you expression.
-     *
-     * @param metricName The metric name.
-     * @param operations The list of operations to create metrics for.
-     * @param props Properties for the individual metrics.
-     * @param metricNameMapper Mapper function to allow controlling the individual metric name per operation.
-     */
-    private createMetricsForOperations(metricName: string, operations: Operation[],
-        props?: cloudwatch.MetricOptions, metricNameMapper?: (op: Operation) => string): Record<string, cloudwatch.IMetric> {
-
-        const metrics: Record<string, cloudwatch.IMetric> = {};
-
-        const mapper = metricNameMapper ?? (op => op.toLowerCase());
-
-        if (props?.dimensionsMap?.Operation) {
-            throw new Error('Invalid properties. Operation dimension is not supported when calculating operational metrics');
-        }
-
-        for (const operation of operations) {
-
-            const metric = this.metric(metricName, {
-                ...props,
-                dimensionsMap: {
-                    TableName: this.tableName,
-                    Operation: operation,
-                    ...props?.dimensionsMap,
-                },
-            });
-
-            const operationMetricName = mapper(operation);
-            const firstChar = operationMetricName.charAt(0);
-
-            if (firstChar === firstChar.toUpperCase()) {
-                // https://docs.aws.amazon.com/AmazonCloudWatch/latest/monitoring/using-metric-math.html#metric-math-syntax
-                throw new Error(`Mapper generated an illegal operation metric name: ${operationMetricName}. Must start with a lowercase letter`);
-            }
-
-            metrics[operationMetricName] = metric;
-        }
-
-        return metrics;
-    }
-
-    protected abstract get hasIndex(): boolean;
-
-    /**
-     * Adds an IAM policy statement associated with this table to an IAM
-     * principal's policy.
-     * @param grantee The principal (no-op if undefined)
-     * @param opts Options for keyActions, tableActions and streamActions
-     */
-    private combinedGrant(
-        grantee: iam.IGrantable,
-        opts: { keyActions?: string[], tableActions?: string[], streamActions?: string[] },
-    ): iam.Grant {
-        if (this.encryptionKey && opts.keyActions) {
-            this.encryptionKey.grant(grantee, ...opts.keyActions);
-        }
-        if (opts.tableActions) {
-            const resources = [this.tableArn,
-            Lazy.string({ produce: () => this.hasIndex ? `${this.tableArn}/index/*` : Aws.NO_VALUE }),
-            ...this.regionalArns,
-            ...this.regionalArns.map(arn => Lazy.string({
-                produce: () => this.hasIndex ? `${arn}/index/*` : Aws.NO_VALUE,
-            }))];
-            const ret = iam.Grant.addToPrincipal({
-                grantee,
-                actions: opts.tableActions,
-                resourceArns: resources,
-                scope: this,
-            });
-            return ret;
-        }
-        if (opts.streamActions) {
-            if (!this.tableStreamArn) {
-                throw new Error(`DynamoDB Streams must be enabled on the table ${this.node.path}`);
-            }
-            const resources = [this.tableStreamArn];
-            const ret = iam.Grant.addToPrincipal({
-                grantee,
-                actions: opts.streamActions,
-                resourceArns: resources,
-                scope: this,
-            });
-            return ret;
-        }
-        throw new Error(`Unexpected 'action', ${opts.tableActions || opts.streamActions}`);
-    }
-
-    private cannedMetric(
-        fn: (dims: { TableName: string }) => cloudwatch.MetricProps,
-        props?: cloudwatch.MetricOptions): cloudwatch.Metric {
-        return new cloudwatch.Metric({
-            ...fn({ TableName: this.tableName }),
-            ...props,
-        }).attachTo(this);
-    }
-}
-
-/**
- * Provides a DynamoDB table.
+ * Provides a Global DynamoDB table.
  */
 export class GlobalTable extends TableBase {
     /**
@@ -1310,11 +396,6 @@ export class GlobalTable extends TableBase {
     private readonly tableSortKey?: Attribute;
 
     private readonly billingMode: BillingMode;
-    private readonly tableScaling: ScalableAttributePair = {};
-    private readonly indexScaling = new Map<string, ScalableAttributePair>();
-    private readonly scalingRole: iam.IRole;
-
-    private readonly globalReplicaCustomResources = new Array<CustomResource>();
 
     constructor(scope: Construct, id: string, props: TableProps) {
         super(scope, id, {
@@ -1332,14 +413,14 @@ export class GlobalTable extends TableBase {
             }
             streamSpecification = { streamViewType: StreamViewType.NEW_AND_OLD_IMAGES };
 
-            this.billingMode = props.billingMode ?? BillingMode.PAY_PER_REQUEST;
+            this.billingMode = props.billingMode ?? BillingMode.PROVISIONED;
         } else {
             this.billingMode = props.billingMode ?? BillingMode.PAY_PER_REQUEST;
             if (props.stream) {
                 streamSpecification = { streamViewType: props.stream };
             }
         }
-        this.validateProvisioning(props);
+        this.validateProvisioning(props.capacity?.provisionedCapacity);
         this.replicas.push({
             contributorInsightsSpecification: props.contributorInsightsEnabled !== undefined ? { enabled: props.contributorInsightsEnabled } : undefined,
             kinesisStreamSpecification: props.kinesisStream ? { streamArn: props.kinesisStream.streamArn } : undefined,
@@ -1386,8 +467,6 @@ export class GlobalTable extends TableBase {
 
         this.tableStreamArn = streamSpecification ? this.globalTable.attrStreamArn : undefined;
 
-        this.scalingRole = this.makeScalingRole();
-
         this.addKey(props.partitionKey, HASH_KEY_TYPE);
         this.tablePartitionKey = props.partitionKey;
 
@@ -1396,8 +475,10 @@ export class GlobalTable extends TableBase {
             this.tableSortKey = props.sortKey;
         }
 
-        if (props.replicationRegions && props.replicationRegions.length > 0) {
-            this.createReplicaTables(props.replicationRegions, props.replicationTimeout, props.waitForReplicationToFinish);
+        if (props.replicas && props.replicas.length > 0) {
+            props.replicas.map((replica: ReplicaProps) => {
+                this.addReplica(replica);
+            })
         }
 
         this.node.addValidation({ validate: () => this.validateTable() });
@@ -1409,9 +490,15 @@ export class GlobalTable extends TableBase {
      * @param props the property of replica table.
      */
     public addReplica(props: ReplicaProps) {
+        const stack = Stack.of(this);
+
+        if (!Token.isUnresolved(stack.region) && props.region.includes(stack.region)) {
+            throw new Error('`replicationRegions` cannot include the region where this stack is deployed.');
+        }
+
         this.replicaGlobalSecondaryIndexes.map((rep: any) => {
-            if(rep.indexName === props.replicaGlobalSecondaryIndexSpecification?.indexName) {
-                rep.readProvisionedThroughputSettings = props.replicaGlobalSecondaryIndexSpecification?.readProvisionedThroughputSettings
+            if (rep.indexName === props.provisionedThroughput?.explicitProvisons?.globalSecondaryIndexes?.indexName) {
+                rep.readProvisionedThroughputSettings = props.provisionedThroughput?.explicitProvisons?.globalSecondaryIndexes?.indexName
             }
         })
         this.replicas.push({
@@ -1420,7 +507,9 @@ export class GlobalTable extends TableBase {
             pointInTimeRecoverySpecification: props.pointInTimeRecovery != null ? { pointInTimeRecoveryEnabled: props.pointInTimeRecovery } : undefined,
             region: props.region,
             tableClass: props.tableClass,
-            readProvisionedThroughputSettings: props.readProvisionedThroughputSettings,
+            readProvisionedThroughputSettings: {
+                readCapacityAutoScalingSettings: props.provisionedThroughput?.explicitProvisons?.table
+            },
             globalSecondaryIndexes: this.replicaGlobalSecondaryIndexes,
         })
     }
@@ -1432,7 +521,7 @@ export class GlobalTable extends TableBase {
      * @param props the property of global secondary index
      */
     public addGlobalSecondaryIndex(props: GlobalSecondaryIndexProps) {
-        this.validateProvisioning(props);
+        this.validateProvisioning(props.capacity?.provisionedCapacity);
         this.validateIndexName(props.indexName);
 
         // build key schema and projection for index
@@ -1443,19 +532,21 @@ export class GlobalTable extends TableBase {
             indexName: props.indexName,
             keySchema: gsiKeySchema,
             projection: gsiProjection,
-            writeProvisionedThroughputSettings: props.writeProvisionedThroughputSettings,
+            writeProvisionedThroughputSettings: {
+                writeCapacityAutoScalingSettings: props.capacity?.provisionedCapacity?.autoScaleWrite
+            },
         });
         this.replicaGlobalSecondaryIndexes.push({
             indexName: props.indexName,
-            readProvisionedThroughputSettings: props.readProvisionedThroughputSettings
+            readProvisionedThroughputSettings: {
+
+            }
         })
 
         this.secondaryIndexSchemas.set(props.indexName, {
             partitionKey: props.partitionKey,
             sortKey: props.sortKey,
         });
-
-        this.indexScaling.set(props.indexName, {});
     }
 
     /**
@@ -1484,106 +575,6 @@ export class GlobalTable extends TableBase {
         this.secondaryIndexSchemas.set(props.indexName, {
             partitionKey: this.tablePartitionKey,
             sortKey: props.sortKey,
-        });
-    }
-
-    /**
-     * Enable read capacity scaling for this table
-     *
-     * @returns An object to configure additional AutoScaling settings
-     */
-    public autoScaleReadCapacity(props: EnableScalingProps): IScalableTableAttribute {
-        if (this.tableScaling.scalableReadAttribute) {
-            throw new Error('Read AutoScaling already enabled for this table');
-        }
-        if (this.billingMode === BillingMode.PAY_PER_REQUEST) {
-            throw new Error('AutoScaling is not available for tables with PAY_PER_REQUEST billing mode');
-        }
-
-        return this.tableScaling.scalableReadAttribute = new ScalableTableAttribute(this, 'ReadScaling', {
-            serviceNamespace: appscaling.ServiceNamespace.DYNAMODB,
-            resourceId: `table/${this.tableName}`,
-            dimension: 'dynamodb:table:ReadCapacityUnits',
-            role: this.scalingRole,
-            ...props,
-        });
-    }
-
-    /**
-     * Enable write capacity scaling for this table
-     *
-     * @returns An object to configure additional AutoScaling settings for this attribute
-     */
-    public autoScaleWriteCapacity(props: EnableScalingProps): IScalableTableAttribute {
-        if (this.tableScaling.scalableWriteAttribute) {
-            throw new Error('Write AutoScaling already enabled for this table');
-        }
-        if (this.billingMode === BillingMode.PAY_PER_REQUEST) {
-            throw new Error('AutoScaling is not available for tables with PAY_PER_REQUEST billing mode');
-        }
-
-        this.tableScaling.scalableWriteAttribute = new ScalableTableAttribute(this, 'WriteScaling', {
-            serviceNamespace: appscaling.ServiceNamespace.DYNAMODB,
-            resourceId: `table/${this.tableName}`,
-            dimension: 'dynamodb:table:WriteCapacityUnits',
-            role: this.scalingRole,
-            ...props,
-        });
-        for (const globalReplicaCustomResource of this.globalReplicaCustomResources) {
-            globalReplicaCustomResource.node.addDependency(this.tableScaling.scalableWriteAttribute);
-        }
-        return this.tableScaling.scalableWriteAttribute;
-    }
-
-    /**
-     * Enable read capacity scaling for the given GSI
-     *
-     * @returns An object to configure additional AutoScaling settings for this attribute
-     */
-    public autoScaleGlobalSecondaryIndexReadCapacity(indexName: string, props: EnableScalingProps): IScalableTableAttribute {
-        if (this.billingMode === BillingMode.PAY_PER_REQUEST) {
-            throw new Error('AutoScaling is not available for tables with PAY_PER_REQUEST billing mode');
-        }
-        const attributePair = this.indexScaling.get(indexName);
-        if (!attributePair) {
-            throw new Error(`No global secondary index with name ${indexName}`);
-        }
-        if (attributePair.scalableReadAttribute) {
-            throw new Error('Read AutoScaling already enabled for this index');
-        }
-
-        return attributePair.scalableReadAttribute = new ScalableTableAttribute(this, `${indexName}ReadScaling`, {
-            serviceNamespace: appscaling.ServiceNamespace.DYNAMODB,
-            resourceId: `table/${this.tableName}/index/${indexName}`,
-            dimension: 'dynamodb:index:ReadCapacityUnits',
-            role: this.scalingRole,
-            ...props,
-        });
-    }
-
-    /**
-     * Enable write capacity scaling for the given GSI
-     *
-     * @returns An object to configure additional AutoScaling settings for this attribute
-     */
-    public autoScaleGlobalSecondaryIndexWriteCapacity(indexName: string, props: EnableScalingProps): IScalableTableAttribute {
-        if (this.billingMode === BillingMode.PAY_PER_REQUEST) {
-            throw new Error('AutoScaling is not available for tables with PAY_PER_REQUEST billing mode');
-        }
-        const attributePair = this.indexScaling.get(indexName);
-        if (!attributePair) {
-            throw new Error(`No global secondary index with name ${indexName}`);
-        }
-        if (attributePair.scalableWriteAttribute) {
-            throw new Error('Write AutoScaling already enabled for this index');
-        }
-
-        return attributePair.scalableWriteAttribute = new ScalableTableAttribute(this, `${indexName}WriteScaling`, {
-            serviceNamespace: appscaling.ServiceNamespace.DYNAMODB,
-            resourceId: `table/${this.tableName}/index/${indexName}`,
-            dimension: 'dynamodb:index:WriteCapacityUnits',
-            role: this.scalingRole,
-            ...props,
         });
     }
 
@@ -1621,28 +612,17 @@ export class GlobalTable extends TableBase {
             errors.push('a sort key of the table must be specified to add local secondary indexes');
         }
 
-        if (this.globalReplicaCustomResources.length > 0 && this.billingMode === BillingMode.PROVISIONED) {
-            const writeAutoScaleAttribute = this.tableScaling.scalableWriteAttribute;
-            if (!writeAutoScaleAttribute) {
-                errors.push('A global Table that uses PROVISIONED as the billing mode needs auto-scaled write capacity. ' +
-                    'Use the autoScaleWriteCapacity() method to enable it.');
-            } else if (!writeAutoScaleAttribute._scalingPolicyCreated) {
-                errors.push('A global Table that uses PROVISIONED as the billing mode needs auto-scaled write capacity with a policy. ' +
-                    'Call one of the scaleOn*() methods of the object returned from autoScaleWriteCapacity()');
-            }
-        }
-
         return errors;
     }
 
     /**
      * Validate read and write capacity are not specified for on-demand tables (billing mode PAY_PER_REQUEST).
      *
-     * @param props read and write capacity properties
+     * @param props read, autoScaleRead and autoScaleWrite capacity properties
      */
-    private validateProvisioning(props: { readCapacity?: number, writeCapacity?: number }): void {
+    private validateProvisioning(props?: provisionedProps): void {
         if (this.billingMode === BillingMode.PAY_PER_REQUEST) {
-            if (props.readCapacity !== undefined || props.writeCapacity !== undefined) {
+            if (props?.read !== undefined || props?.autoScaleRead !== undefined || props?.autoScaleWrite) {
                 throw new Error('you cannot provision read and write capacity for a table with PAY_PER_REQUEST billing mode');
             }
         }
@@ -1747,118 +727,6 @@ export class GlobalTable extends TableBase {
     }
 
     /**
-     * Return the role that will be used for AutoScaling
-     */
-    private makeScalingRole(): iam.IRole {
-        // Use a Service Linked Role.
-        // https://docs.aws.amazon.com/autoscaling/application/userguide/application-auto-scaling-service-linked-roles.html
-        return iam.Role.fromRoleArn(this, 'ScalingRole', Stack.of(this).formatArn({
-            service: 'iam',
-            region: '',
-            resource: 'role/aws-service-role/dynamodb.application-autoscaling.amazonaws.com',
-            resourceName: 'AWSServiceRoleForApplicationAutoScaling_DynamoDBTable',
-        }));
-    }
-
-    /**
-     * Creates replica tables
-     *
-     * @param regions regions where to create tables
-     */
-    private createReplicaTables(regions: string[], timeout?: Duration, waitForReplicationToFinish?: boolean) {
-        const stack = Stack.of(this);
-
-        if (!Token.isUnresolved(stack.region) && regions.includes(stack.region)) {
-            throw new Error('`replicationRegions` cannot include the region where this stack is deployed.');
-        }
-
-        const provider = ReplicaProvider.getOrCreate(this, { tableName: this.tableName, regions, timeout });
-
-        // Documentation at https://docs.aws.amazon.com/amazondynamodb/latest/developerguide/V2gt_IAM.html
-        // is currently incorrect. AWS Support recommends `dynamodb:*` in both source and destination regions
-
-        const onEventHandlerPolicy = new SourceTableAttachedPolicy(this, provider.onEventHandler.role!);
-        const isCompleteHandlerPolicy = new SourceTableAttachedPolicy(this, provider.isCompleteHandler.role!);
-
-        // Permissions in the source region
-        this.grant(onEventHandlerPolicy, 'dynamodb:*');
-        this.grant(isCompleteHandlerPolicy, 'dynamodb:DescribeTable');
-
-        let previousRegion: CustomResource | undefined;
-        let previousRegionCondition: CfnCondition | undefined;
-        for (const region of new Set(regions)) { // Remove duplicates
-            // Use multiple custom resources because multiple create/delete
-            // updates cannot be combined in a single API call.
-            const currentRegion = new CustomResource(this, `Replica${region}`, {
-                serviceToken: provider.provider.serviceToken,
-                resourceType: 'Custom::DynamoDBReplica',
-                properties: {
-                    TableName: this.tableName,
-                    Region: region,
-                    SkipReplicationCompletedWait: waitForReplicationToFinish == null
-                        ? undefined
-                        // CFN changes Custom Resource properties to strings anyways,
-                        // so let's do that ourselves to make it clear in the handler this is a string, not a boolean
-                        : (!waitForReplicationToFinish).toString(),
-                },
-            });
-            currentRegion.node.addDependency(
-                onEventHandlerPolicy.policy,
-                isCompleteHandlerPolicy.policy,
-            );
-            this.globalReplicaCustomResources.push(currentRegion);
-
-            // Deploy time check to prevent from creating a replica in the region
-            // where this stack is deployed. Only needed for environment agnostic
-            // stacks.
-            let createReplica: CfnCondition | undefined;
-            if (Token.isUnresolved(stack.region)) {
-                createReplica = new CfnCondition(this, `StackRegionNotEquals${region}`, {
-                    expression: Fn.conditionNot(Fn.conditionEquals(region, Aws.REGION)),
-                });
-                const cfnCustomResource = currentRegion.node.defaultChild as CfnCustomResource;
-                cfnCustomResource.cfnOptions.condition = createReplica;
-            }
-
-            // Save regional arns for grantXxx() methods
-            this.regionalArns.push(stack.formatArn({
-                region,
-                service: 'dynamodb',
-                resource: 'table',
-                resourceName: this.tableName,
-            }));
-
-            // We need to create/delete regions sequentially because we cannot
-            // have multiple table updates at the same time. The `isCompleteHandler`
-            // of the provider waits until the replica is in an ACTIVE state.
-            if (previousRegion) {
-                if (previousRegionCondition) {
-                    // we can't simply use a Dependency,
-                    // because the previousRegion is protected by the "different region" Condition,
-                    // and you can't have Fn::If in DependsOn.
-                    // Instead, rely on Ref adding a dependency implicitly!
-                    const previousRegionCfnResource = previousRegion.node.defaultChild as CfnResource;
-                    const currentRegionCfnResource = currentRegion.node.defaultChild as CfnResource;
-                    currentRegionCfnResource.addMetadata('DynamoDbReplicationDependency',
-                        Fn.conditionIf(previousRegionCondition.logicalId, previousRegionCfnResource.ref, Aws.NO_VALUE));
-                } else {
-                    currentRegion.node.addDependency(previousRegion);
-                }
-            }
-
-            previousRegion = currentRegion;
-            previousRegionCondition = createReplica;
-        }
-
-        // Permissions in the destination regions (outside of the loop to
-        // minimize statements in the policy)
-        onEventHandlerPolicy.grantPrincipal.addToPrincipalPolicy(new iam.PolicyStatement({
-            actions: ['dynamodb:*'],
-            resources: this.regionalArns,
-        }));
-    }
-
-    /**
      * Whether this table has indexes
      */
     protected get hasIndex(): boolean {
@@ -1892,7 +760,7 @@ export class GlobalTable extends TableBase {
             throw new Error('`encryptionKey cannot be specified unless encryption is set to TableEncryption.CUSTOMER_MANAGED (it was set to ${encryptionType})`');
         }
 
-        if (encryptionType === TableEncryption.CUSTOMER_MANAGED && props.replicationRegions) {
+        if (encryptionType === TableEncryption.CUSTOMER_MANAGED && props.replicas) {
             throw new Error('TableEncryption.CUSTOMER_MANAGED is not supported by DynamoDB Global Tables (where replicationRegions was set)');
         }
 
@@ -1925,164 +793,6 @@ export class GlobalTable extends TableBase {
     }
 }
 
-/**
- * Data types for attributes within a table
- *
- * @see https://docs.aws.amazon.com/amazondynamodb/latest/developerguide/HowItWorks.NamingRulesDataTypes.html#HowItWorks.DataTypes
- */
-export enum AttributeType {
-    /** Up to 400KiB of binary data (which must be encoded as base64 before sending to DynamoDB) */
-    BINARY = 'B',
-    /** Numeric values made of up to 38 digits (positive, negative or zero) */
-    NUMBER = 'N',
-    /** Up to 400KiB of UTF-8 encoded text */
-    STRING = 'S',
-}
-
-/**
- * DynamoDB's Read/Write capacity modes.
- */
-export enum BillingMode {
-    /**
-     * Pay only for what you use. You don't configure Read/Write capacity units.
-     */
-    PAY_PER_REQUEST = 'PAY_PER_REQUEST',
-    /**
-     * Explicitly specified Read/Write capacity units.
-     */
-    PROVISIONED = 'PROVISIONED',
-}
-
-/**
- * The set of attributes that are projected into the index
- *
- * @see https://docs.aws.amazon.com/amazondynamodb/latest/APIReference/API_Projection.html
- */
-export enum ProjectionType {
-    /** Only the index and primary keys are projected into the index. */
-    KEYS_ONLY = 'KEYS_ONLY',
-    /** Only the specified table attributes are projected into the index. The list of projected attributes is in `nonKeyAttributes`. */
-    INCLUDE = 'INCLUDE',
-    /** All of the table attributes are projected into the index. */
-    ALL = 'ALL'
-}
-
-/**
- * When an item in the table is modified, StreamViewType determines what information
- * is written to the stream for this table.
- *
- * @see https://docs.aws.amazon.com/amazondynamodb/latest/APIReference/API_StreamSpecification.html
- */
-export enum StreamViewType {
-    /** The entire item, as it appears after it was modified, is written to the stream. */
-    NEW_IMAGE = 'NEW_IMAGE',
-    /** The entire item, as it appeared before it was modified, is written to the stream. */
-    OLD_IMAGE = 'OLD_IMAGE',
-    /** Both the new and the old item images of the item are written to the stream. */
-    NEW_AND_OLD_IMAGES = 'NEW_AND_OLD_IMAGES',
-    /** Only the key attributes of the modified item are written to the stream. */
-    KEYS_ONLY = 'KEYS_ONLY'
-}
-
-/**
- * DynamoDB's table class.
- *
- * @see https://docs.aws.amazon.com/amazondynamodb/latest/developerguide/HowItWorks.TableClasses.html
- */
-export enum TableClass {
-    /** Default table class for DynamoDB. */
-    STANDARD = 'STANDARD',
-
-    /** Table class for DynamoDB that reduces storage costs compared to existing DynamoDB Standard tables. */
-    STANDARD_INFREQUENT_ACCESS = 'STANDARD_INFREQUENT_ACCESS',
-}
-
-/**
- * Just a convenient way to keep track of both attributes
- */
-interface ScalableAttributePair {
-    scalableReadAttribute?: ScalableTableAttribute;
-    scalableWriteAttribute?: ScalableTableAttribute;
-}
-
-/**
- * An inline policy that is logically bound to the source table of a DynamoDB Global Tables
- * "cluster". This is here to ensure permissions are removed as part of (and not before) the
- * CleanUp phase of a stack update, when a replica is removed (or the entire "cluster" gets
- * replaced).
- *
- * If statements are added directly to the handler roles (as opposed to in a separate inline
- * policy resource), new permissions are in effect before clean up happens, and so replicas that
- * need to be dropped can no longer be due to lack of permissions.
- */
-class SourceTableAttachedPolicy extends Construct implements iam.IGrantable {
-    public readonly grantPrincipal: iam.IPrincipal;
-    public readonly policy: iam.IManagedPolicy;
-
-    public constructor(sourceTable: GlobalTable, role: iam.IRole) {
-        super(sourceTable, `SourceTableAttachedManagedPolicy-${Names.nodeUniqueId(role.node)}`);
-
-        const policy = new iam.ManagedPolicy(this, 'Resource', {
-            // A CF update of the description property of a managed policy requires
-            // a replacement. Use the table name in the description to force a managed
-            // policy replacement when the table name changes. This way we preserve permissions
-            // to delete old replicas in case of a table replacement.
-            description: `DynamoDB replication managed policy for table ${sourceTable.tableName}`,
-            roles: [role],
-        });
-        this.policy = policy;
-        this.grantPrincipal = new SourceTableAttachedPrincipal(role, policy);
-    }
-}
-
-/**
- * An `IPrincipal` entity that can be used as the target of `grant` calls, used by the
- * `SourceTableAttachedPolicy` class so it can act as an `IGrantable`.
- */
-class SourceTableAttachedPrincipal extends iam.PrincipalBase {
-    public constructor(private readonly role: iam.IRole, private readonly policy: iam.ManagedPolicy) {
-        super();
-    }
-
-    public get policyFragment(): iam.PrincipalPolicyFragment {
-        return this.role.policyFragment;
-    }
-
-    public addToPrincipalPolicy(statement: iam.PolicyStatement): iam.AddToPrincipalPolicyResult {
-        this.policy.addStatements(statement);
-        return {
-            policyDependable: this.policy,
-            statementAdded: true,
-        };
-    }
-
-    public dedupeString(): string | undefined {
-        return undefined;
-    }
-}
-
-// interface pp {
-//     /**
-//      * Minimum capacity to scale to
-//      */
-//     readonly minCapacity: number;
-
-//     /**
-//      * Maximum capacity to scale to
-//      */
-//     readonly maxCapacity: number;
-
-//     /**
-//     * Add scheduled scaling for this scaling attribute
-//     */
-//     scaleOnSchedule(id: string, actions: appscaling.ScalingSchedule): void;
-
-//     /**
-//      * Scale out or in to keep utilization at a given level
-//      */
-//     scaleOnUtilization(props: UtilizationScalingProps): void;
-// }
-
 interface provisionedProps {
     read?: number,
     // write?: number,
@@ -2103,7 +813,7 @@ export class Capacity {
     constructor(billingMode: BillingMode, props?: provisionedProps) {
         this.billMode = billingMode
         this.provisionedCapacity = props
-     }
+    }
 
     //public readonly autoR?: IScalableTableAttribute
     //public readonly autoW?: IScalableTableAttribute
@@ -2118,16 +828,16 @@ export class Capacity {
 }
 
 interface explicitProps {
-    table: CfnGlobalTable.CapacityAutoScalingSettingsProperty | number
+    table: CfnGlobalTable.CapacityAutoScalingSettingsProperty
     globalSecondaryIndexes?: {
-        [indexname: string]: CfnGlobalTable.CapacityAutoScalingSettingsProperty | number
+        [indexname: string]: CfnGlobalTable.CapacityAutoScalingSettingsProperty
     }
 }
 
-interface ThroughputProvisions {
-    autoScale?: CfnGlobalTable.CapacityAutoScalingSettingsProperty
-    read?: number
-}
+// interface ThroughputProvisions {
+//     autoScale?: CfnGlobalTable.CapacityAutoScalingSettingsProperty
+//     read?: number
+// }
 
 interface ThroughputProps {
     explicitProvisons?: explicitProps
